LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/parameter-types.tentative-expected.txt


FAIL A parameter retains its type assert_equals: expected "PASS" but got "if(\n        style(--c:red): PASS;\n        else: FAIL)"
FAIL A parameter type acts as a local registration assert_equals: expected "PASS" but got "if(\n        style(--c:blue): PASS;\n        else: FAIL)"
FAIL A parameter retains its type (parent stack frame) assert_equals: expected "PASS" but got "--f()"
FAIL A parameter type acts as a local registration (parent stack frame) assert_equals: expected "PASS" but got "--f()"
FAIL Universally typed parameter can shadow other parameters assert_equals: expected "PASS" but got "--f(#0f0)"
FAIL Invalid value for typed local becomes IACVT assert_equals: expected "PASS" but got "var(--c, PASS)"
FAIL if() within @function can query registered custom property assert_equals: expected "PASS" but got "if(style(--c:#f00): PASS; else: FAIL)"
